+++ packages/react-storage/src/components/FileUploader/hooks/useFileUploader/useFileUploader.ts
@@ -10,10 +10,10 @@ import {
   clearFilesAction,
   queueFilesAction,
   removeUploadAction,
-  setProcessedKeyAction,
   setUploadingFileAction,
   setUploadProgressAction,
   setUploadStatusAction,
+  setUploadSuccessAction,
 } from './actions';
 import { TaskHandler } from '../../utils';
 
@@ -24,15 +24,14 @@ export interface UseFileUploader {
     getFileErrorMessage: GetFileErrorMessage;
   }) => void;
   clearFiles: () => void;
-  queueFiles: () => void;
-  setUploadingFile: TaskHandler;
-  setProcessedKey: (params: { id: string; processedKey: string }) => void;
-  setUploadProgress: (params: { id: string; progress: number }) => void;
-  setUploadSuccess: (params: { id: string }) => void;
-  setUploadResumed: (params: { id: string }) => void;
-  setUploadPaused: (params: { id: string }) => void;
-  removeUpload: (params: { id: string }) => void;
   files: StorageFiles;
+  queueFiles: () => void;
+  removeUpload: (params: { id: string }) => void;
+  setUploadingFile: TaskHandler;
+  setUploadPaused: (params: { id: string }) => void;
+  setUploadProgress: (params: { id: string; progress: number }) => void;
+  setUploadResumed: (params: { id: string }) => void;
+  setUploadSuccess: (params: { id: string; resolvedKey: string }) => void;
 }
 
 const isDefaultFile = (file: unknown): file is DefaultFile =>
@@ -54,67 +53,38 @@ export function useFileUploader(
       : []) as StorageFiles,
   });
 
-  const addFiles: UseFileUploader['addFiles'] = ({
-    files,
-    status,
-    getFileErrorMessage,
-  }) => {
-    dispatch(addFilesAction({ files, status, getFileErrorMessage }));
-  };
+  const dispatchers: Omit<UseFileUploader, 'files'> = React.useMemo(
+    () => ({
+      addFiles: (params) => {
+        dispatch(addFilesAction(params));
+      },
+      clearFiles: () => {
+        dispatch(clearFilesAction());
+      },
+      queueFiles: () => {
+        dispatch(queueFilesAction());
+      },
+      setUploadingFile: (params) => {
+        dispatch(setUploadingFileAction(params));
+      },
+      setUploadProgress: (params) => {
+        dispatch(setUploadProgressAction(params));
+      },
+      setUploadSuccess: (params) => {
+        dispatch(setUploadSuccessAction(params));
+      },
+      setUploadPaused: ({ id }) => {
+        dispatch(setUploadStatusAction({ id, status: FileStatus.PAUSED }));
+      },
+      setUploadResumed: ({ id }) => {
+        dispatch(setUploadStatusAction({ id, status: FileStatus.UPLOADING }));
+      },
+      removeUpload: ({ id }) => {
+        dispatch(removeUploadAction({ id }));
+      },
+    }),
+    []
+  );
 
-  const clearFiles: UseFileUploader['clearFiles'] = () => {
-    dispatch(clearFilesAction());
-  };
-
-  const queueFiles: UseFileUploader['queueFiles'] = () => {
-    dispatch(queueFilesAction());
-  };
-
-  const setUploadingFile: UseFileUploader['setUploadingFile'] = ({
-    uploadTask,
-    id,
-  }) => {
-    dispatch(setUploadingFileAction({ id, uploadTask }));
-  };
-
-  const setProcessedKey: UseFileUploader['setProcessedKey'] = (input) => {
-    dispatch(setProcessedKeyAction(input));
-  };
-
-  const setUploadProgress: UseFileUploader['setUploadProgress'] = ({
-    progress,
-    id,
-  }) => {
-    dispatch(setUploadProgressAction({ id, progress }));
-  };
-
-  const setUploadSuccess: UseFileUploader['setUploadSuccess'] = ({ id }) => {
-    dispatch(setUploadStatusAction({ id, status: FileStatus.UPLOADED }));
-  };
-
-  const setUploadPaused: UseFileUploader['setUploadPaused'] = ({ id }) => {
-    dispatch(setUploadStatusAction({ id, status: FileStatus.PAUSED }));
-  };
-
-  const setUploadResumed: UseFileUploader['setUploadPaused'] = ({ id }) => {
-    dispatch(setUploadStatusAction({ id, status: FileStatus.UPLOADING }));
-  };
-
-  const removeUpload: UseFileUploader['removeUpload'] = ({ id }) => {
-    dispatch(removeUploadAction({ id }));
-  };
-
-  return {
-    removeUpload,
-    setProcessedKey,
-    setUploadPaused,
-    setUploadProgress,
-    setUploadResumed,
-    setUploadSuccess,
-    setUploadingFile,
-    queueFiles,
-    addFiles,
-    clearFiles,
-    files,
-  };
+  return { ...dispatchers, files };
 }