+++ packages/react-storage/src/components/FileUploader/types.ts
@@ -26,8 +26,8 @@ export interface StorageFile {
   file?: File;
   status: FileStatus;
   progress: number;
-  // only present after `processFile` complete
-  processedKey?: string;
+  // only present after an upload completes
+  resolvedKey?: string;
   uploadTask?: UploadTask;
   key: string;
   error: string;