+++ packages/react-storage/src/components/StorageManager/ui/FileList/FileStatusMessage.tsx
@@ -5,7 +5,7 @@ import { ComponentClassName } from '@aws-amplify/ui';
 import { Text, View } from '@aws-amplify/ui-react';
 import { IconCheck, IconError, useIcons } from '@aws-amplify/ui-react/internal';
 import { classNameModifier } from '@aws-amplify/ui';
-import { FileStatus } from '../../types';
+import { FileStatus } from '../../../FileUploader/types';
 import { FileStatusMessageProps } from './types';
 
 export const FileStatusMessage = ({